tests: put filechooser and setInputFiles tests into separate files
